Add files via upload
Optimal Liquidation problem: how to optimally liquidate a certain number of shares ?
Combination of Reinforcement Learning and Algorithmic trading. The agent has to find the optimal rate to sell all the shares in a 2 hour window.
The goal is to handle the trade-off between price risk and price impact, using the predictive power of the order-flow.